Refactor code structure and relocate functions
Split the RomanNumerals function into distinct components for clarity and better organization. RomanNumeralExtension is now in Solution, while RomanNumeralsFunction is moved under Function namespace. Also, several other functions have been renamed and moved into the Function namespace.
